RAG/requirements.txt
langchain>=0.3
markdown-it-py>=3.0
langchain-community>=0.3
flashrank>=0.2
langchain-ollama>=0.3
langchain-text-splitters>=0.3
pydantic>=2.7
pydantic-settings>=2.5
python-dotenv>=1.0
clickhouse-connect>=0.7

